fix: silent printing default dpi on Windows
